+++ crates/sui-json-rpc/Cargo.toml
@@ -26,7 +26,6 @@ sui-core = { path = "../sui-core" }
 sui-types = { path = "../sui-types" }
 sui-json = { path = "../sui-json" }
 sui-open-rpc = { path = "../sui-open-rpc" }
-sui-quorum-driver = { path = "../sui-quorum-driver" }
 sui-open-rpc-macros = { path = "../sui-open-rpc-macros" }
 sui-json-rpc-types = { path = "../sui-json-rpc-types" }
 sui-cost = { path = "../sui-cost" }
