+++ crates/sui-json-rpc/src/quorum_driver_api.rs
@@ -12,9 +12,9 @@ use signature::Signature;
 use std::sync::Arc;
 use sui_core::authority::{AuthorityStore, ResolverWrapper};
 use sui_core::authority_client::NetworkAuthorityClient;
+use sui_core::quorum_driver::QuorumDriver;
 use sui_json_rpc_types::SuiExecuteTransactionResponse;
 use sui_open_rpc::Module;
-use sui_quorum_driver::QuorumDriver;
 use sui_types::crypto::SignatureScheme;
 use sui_types::messages::{ExecuteTransactionRequest, ExecuteTransactionRequestType};
 use sui_types::sui_serde::Base64;
