--- crates/sui-quorum-driver/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "sui-quorum-driver"
-version = "0.1.0"
-authors = ["Mysten Labs <build@mystenlabs.com>"]
-license = "Apache-2.0"
-publish = false
-edition = "2021"
-
-[dependencies]
-arc-swap = "1.5.1"
-tokio = { version = "1.20.1", features = ["full"] }
-tracing = "0.1.36"
-prometheus = "0.13.2"
-scopeguard = "1.1"
-
-sui-core = { path = "../sui-core" }
-sui-types = { path = "../sui-types" }
-workspace-hack = { path = "../workspace-hack"}
